Add requirements.txt, show_pcd.py, and update README
This commit adds a `requirements.txt` file that lists the required Python packages for the project. It also adds the `show_pcd.py` script, which can be used to visualize a point cloud map of the environment. Finally, it updates the README to include download links for the public and private sets of the ITRI dataset, as well as instructions for how to use the dataset and the `show_pcd.py` script. Overall, this commit should make it easier for users to get started with the project and work with the dataset.